+++ transaction/_manager.py
@@ -144,7 +144,10 @@ class TransactionManager(object):
         while number:
             number -= 1
             if number:
-                yield Attempt(self)
+                attempt = Attempt(self)
+                yield attempt
+                if attempt.success:
+                    break
             else:
                 yield self
 
@@ -167,6 +170,8 @@ class ThreadTransactionManager(TransactionManager, threading.local):
 
 class Attempt(object):
 
+    success = False
+
     def __init__(self, manager):
         self.manager = manager
 
@@ -186,5 +191,7 @@ class Attempt(object):
                 self.manager.commit()
             except:
                 return self._retry_or_raise(*sys.exc_info())
+            else:
+                self.success = True
         else:
             return self._retry_or_raise(t, v, tb)